+++ database/connector/dialect/sqlserver/src/main/java/org/apache/shardingsphere/database/connector/sql92/sqlserver/metadata/database/SQLServerDatabaseMetaData.java
@@ -89,6 +89,11 @@ public final class SQLServerDatabaseMetaData implements DialectDatabaseMetaData 
     }
     
     @Override
+    public Optional<String> getVariableTableNamePrefix() {
+        return Optional.of("@");
+    }
+    
+    @Override
     public String getDatabaseType() {
         return "SQLServer";
     }